requirements/requirements-dev.txt

# appdev requirements

django-debug-toolbar
django-haystack-panel

Werkzeug

#requests
#requests-oauthlib
#httpie
#httpie-oauth # OAuth1
curlish

-r requirements-test.txt
